Rename sympy_backends -> sympy_backend
